rename StateContextProvider to AppController; rename zustand states folder to `stores` and move out of contexts
