+++ src/components/DropFileContainer/index.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 
-import useLogFileStore from "../../contexts/states/logFileStore";
-import useUiStore from "../../contexts/states/uiStore";
+import useLogFileStore from "../../stores/logFileStore";
+import useUiStore from "../../stores/uiStore";
 import {UI_ELEMENT} from "../../typings/states";
 import {CURSOR_CODE} from "../../typings/worker";
 import {isDisabled} from "../../utils/states";
