+++ src/components/MenuBar/NavigationBar.tsx
@@ -7,8 +7,8 @@ import NavigateNext from "@mui/icons-material/NavigateNext";
 import SkipNext from "@mui/icons-material/SkipNext";
 import SkipPrevious from "@mui/icons-material/SkipPrevious";
 
-import useUiStore from "../../contexts/states/uiStore";
-import useViewStore from "../../contexts/states/viewStore";
+import useUiStore from "../../stores/uiStore";
+import useViewStore from "../../stores/viewStore";
 import {UI_ELEMENT} from "../../typings/states";
 import {ACTION_NAME} from "../../utils/actions";
 import {
