+++ src/stores/logExportStore.ts
@@ -1,10 +1,10 @@
 import {create} from "zustand";
 
-import LogExportManager, {EXPORT_LOGS_PROGRESS_VALUE_MIN} from "../../services/LogExportManager";
-import {Nullable} from "../../typings/common";
-import {LOG_LEVEL} from "../../typings/logs";
-import {DO_NOT_TIMEOUT_VALUE} from "../../typings/notifications";
-import {EXPORT_LOGS_CHUNK_SIZE} from "../../utils/config";
+import LogExportManager, {EXPORT_LOGS_PROGRESS_VALUE_MIN} from "../services/LogExportManager";
+import {Nullable} from "../typings/common";
+import {LOG_LEVEL} from "../typings/logs";
+import {DO_NOT_TIMEOUT_VALUE} from "../typings/notifications";
+import {EXPORT_LOGS_CHUNK_SIZE} from "../utils/config";
 import useContextStore from "./contextStore";
 import useLogFileManagerProxyStore from "./logFileManagerProxyStore";
 import useLogFileStore from "./logFileStore";
