+++ src/stores/queryStore/queryConfigSlice.ts
